Omnibox is working, but not testing
