+++ js/jasmine/SpecRunner.html
@@ -19,10 +19,10 @@
 
     <!-- include source files here... -->
     <script src="spec/mockChrome.js"></script>
-    <script src="../popup.js"></script>
+    <script src="../main.js"></script>
 
     <!-- include spec files here... -->
-    <script src="spec/PopupSpec.js"></script>
+    <script src="spec/MainSpec.js"></script>
   </head>
 
   <body></body>
