+++ manifest.json
@@ -2,7 +2,7 @@
     "name": "Open JIRA Ticket",
     "description": "Open JIRA tickets easily using the toolbar.",
     "manifest_version": 3,
-    "version": "3.0.0",
+    "version": "3.0.1",
     "author": "jjwong",
     "commands": {
         "_execute_action": {
@@ -13,6 +13,13 @@
                 "linux": "Ctrl+Shift+K"
             }
         }
+    },
+    "background": {
+        "service_worker": "js/background.js", "type": "module"
+    },
+    "minimum_chrome_version": "92",
+    "omnibox": {
+        "keyword": "ojira"
     },
     "icons": {
         "16": "icons/32-toolbar.png",
